refactor: separate snap and mvt commands
